[OP-295] local date supporting date chooser
